product options controller extracted from product controller
Product service draft created
