+++ refactor-me/Controllers/ProductOptionsController.cs
@@ -1,0 +1,59 @@
+﻿using System;
+using System.Net;
+using System.Web.Http;
+using refactor_me.Models;
+
+namespace refactor_me.Controllers
+{
+    [RoutePrefix("products")]
+    public class ProductOptionsController : ApiController
+    {
+        [Route("{productId}/options")]
+        [HttpGet]
+        public ProductOptions GetOptions(Guid productId)
+        {
+            return new ProductOptions(productId);
+        }
+
+        [Route("{productId}/options/{id}")]
+        [HttpGet]
+        public ProductOption GetOption(Guid productId, Guid id)
+        {
+            var option = new ProductOption(id);
+            if (option.IsNew)
+                throw new HttpResponseException(HttpStatusCode.NotFound);
+
+            return option;
+        }
+
+        [Route("{productId}/options")]
+        [HttpPost]
+        public void CreateOption(Guid productId, ProductOption option)
+        {
+            option.ProductId = productId;
+            option.Save();
+        }
+
+        [Route("{productId}/options/{id}")]
+        [HttpPut]
+        public void UpdateOption(Guid id, ProductOption option)
+        {
+            var orig = new ProductOption(id)
+            {
+                Name = option.Name,
+                Description = option.Description
+            };
+
+            if (!orig.IsNew)
+                orig.Save();
+        }
+
+        [Route("{productId}/options/{id}")]
+        [HttpDelete]
+        public void DeleteOption(Guid id)
+        {
+            var opt = new ProductOption(id);
+            opt.Delete();
+        }
+    }
+}
