+++ refactor-me/Services/ProductService.cs
@@ -1,0 +1,43 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+using refactor_me.Models;
+
+namespace refactor_me.Services
+{
+    public class ProductService
+    {
+        public Products GetAll()
+        {
+            throw new  NotImplementedException();
+        }
+
+        public Products SearchByName(string name)
+        {
+            throw new NotImplementedException();
+        }
+
+        public Product GetProduct(Guid id)
+        {
+            throw new NotImplementedException();
+        }
+
+        public Product Create(Product product)
+        {
+            throw new NotImplementedException();
+        }
+
+        public Product Update(Guid id, Product product)
+        {
+            throw new NotImplementedException();
+        }
+
+        public void Delete(Guid id)
+        {
+            throw new NotImplementedException();
+        }
+
+
+    }
+}
